Created colorController and changed minor things
