+++ MementoMori/MementoMori.Server/Controllers/AuthController.cs
@@ -32,7 +32,6 @@ namespace MementoMori.Server.Controllers
             }
         }
 
-
         [HttpPost("register")]
         public async Task<ActionResult> Register([FromBody] RegisterDetails registerDetails)
         {
@@ -85,17 +84,8 @@ namespace MementoMori.Server.Controllers
         public IActionResult GetLoginResponse()
         {
             var userId = _authService.GetRequesterId(HttpContext);
-            if (!userId.HasValue)
-            {
-                return Ok(new LoginResponse { IsLoggedIn = false });
-            }
-
-            var loginResponse = new LoginResponse
-            {
-                IsLoggedIn = true
-            };
-
-            return Ok(loginResponse);
+            bool isLoggedIn = userId.HasValue;
+            return Ok(isLoggedIn);
         }
 
         [HttpPost("logout")]
@@ -105,48 +95,5 @@ namespace MementoMori.Server.Controllers
             return Ok();
         }
 
-        [HttpGet("color")]
-        public async Task<IActionResult> GetUserColor()
-        {
-            var userId = _authService.GetRequesterId(HttpContext);
-            if (!userId.HasValue)
-            {
-                return BadRequest(new { Message = "User ID not found." });
-            }
-
-            var user = await _authRepo.GetUserByIdAsync(userId.Value);
-            if (user == null)
-            {
-                return NotFound(new { Message = "User not found." });
-            }
-
-            return Ok(new { color = user.HeaderColor });
-        }
-
-        [HttpPost("newColor")]
-        public async Task<IActionResult> UpdateUserColor([FromBody] UpdateColorRequest request)
-        {
-            var userId = _authService.GetRequesterId(HttpContext);
-            if (!userId.HasValue)
-            {
-                return Unauthorized();
-            }
-
-            var user = await _authRepo.GetUserByIdAsync(userId.Value);
-            if (user == null)
-            {
-                return NotFound(new { Message = "User not found." });
-            }
-
-            user.HeaderColor = request.NewColor;
-
-            var success = await _authRepo.SaveAsync();
-            if (!success)
-            {
-                return StatusCode(500, new { Message = "Failed to update user color." });
-            }
-
-            return Ok(new { Message = "Header color updated successfully." });
-        }
     }
 }
